Revert "Merge branch 'MonkeyIsland' of https://github.com/Aurum22/AurumBayFixes into MonkeyIsland"
This reverts commit 5208199d9b9d675fec8d0aec7a37f0416a0488e8, reversing
changes made to e7dff9df7c6c78921abdc4a9b718edb2abe17f6c.

+++ code/modules/integrated_electronics/subtypes/reagents.dm
@@ -374,12 +374,18 @@
 		activate_pin(3)
 		return FALSE
 	var/obj/item/I = get_pin_data_as_type(IC_INPUT, 1, /obj/item)
+
+	if(isnull(I))
+		return FALSE
+
 	if(!I.reagents || !I.reagents.total_volume)
 		activate_pin(3)
 		return FALSE
+	
 	I.reagents.trans_to(src,I.reagents.total_volume)
 	if(!I.reagents.total_volume)
 		qdel(I)
+
 	activate_pin(2)
 	return FALSE
